% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{batch_process_data}
\alias{batch_process_data}
\title{Optimize Session Performance
Batch Process Data}
\usage{
batch_process_data(session, data_list, batch_size = 32)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}

\item{data_list}{A list of input data to process}

\item{batch_size}{Number of items to process in each batch}
}
\value{
A list of results from batch processing
}
\description{
Process data in batches for memory efficiency with large datasets.
}
