% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{estimate_session_memory}
\alias{estimate_session_memory}
\title{Estimate Memory Usage}
\usage{
estimate_session_memory(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
Estimated memory usage in bytes
}
\description{
Estimate the memory usage of an ONNX session.
}
\examples{
\dontrun{
session <- onnx_session("path/to/model.onnx")
memory_usage <- estimate_session_memory(session)
cat("Estimated memory usage:", memory_usage / (1024^2), "MB\n")
}
}
