% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{get_session_performance_stats}
\alias{get_session_performance_stats}
\title{Get Session Performance Statistics}
\usage{
get_session_performance_stats(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
A list containing performance statistics
}
\description{
Retrieve performance statistics and memory usage information for a session.
}
\examples{
\dontrun{
session <- onnx_session("path/to/model.onnx")
stats <- get_session_performance_stats(session)
print(stats)
}
}
