% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_onnxruntime}
\alias{install_onnxruntime}
\title{Install onnxruntime}
\usage{
install_onnxruntime(reinstall = FALSE, ..., .inform_restart = TRUE)
}
\arguments{
\item{reinstall}{Re-install onnxruntime even if its already installed?}

\item{...}{Currently unused.}

\item{.inform_restart}{if \code{TRUE} and running in an \code{interactive()} session, after
installation it will print a message to inform the user that the session must
be restarted for onnxruntime to work correctly.}
}
\description{
Installs onnxruntime and its dependencies.
}
\details{
This function is mainly controlled by environment variables that can be used
to override the defaults:
\itemize{
\item \code{ONNXRUNTIME_HOME}: the installation path. By default dependencies are installed
within the package directory. Eg what's given by \code{system.file(package="churon")}.
\item \code{ONNXRUNTIME_URL}: A URL, path to a ZIP file or a directory containing a Libonnxruntime version.
Files will be installed/copied to the \code{ONNXRUNTIME_HOME} directory.
\item \code{ONNXRUNTIME_INSTALL_DEBUG}: Setting it to 1, shows debug log messages during installation.
\item \code{ONNXRUNTIME_COMMIT_SHA}: onnxruntime repository commit sha to be used when querying lantern
uploads. Set it to \code{'none'} to avoid looking for build for that commit and
use the latest build for the branch.
}

The \code{ONNXRUNTIME_INSTALL} environment
variable can be set to \code{0} to prevent auto-installing onnxruntime and \code{ONNXRUNTIME_LOAD} set to \code{0}
to avoid loading dependencies automatically. These environment variables are meant for advanced use
cases and troubleshooting only.
When timeout error occurs during library archive download, or length of downloaded files differ from
reported length, an increase of the \code{timeout} value should help.
}
