% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_download_models}
\alias{onnx_download_models}
\title{배치로 여러 모델 다운로드}
\usage{
onnx_download_models(model_names, dest_dir = ".", overwrite = FALSE)
}
\arguments{
\item{model_names}{모델 이름들의 벡터}

\item{dest_dir}{다운로드 디렉토리}

\item{overwrite}{기존 파일 덮어쓰기}
}
\value{
다운로드된 파일 경로들의 리스트
}
\description{
ONNX Model Zoo에서 여러 모델을 배치로 다운로드합니다.
}
\examples{
\dontrun{
# 여러 모델 다운로드
paths <- onnx_download_models(c("mnist", "squeezenet1.0"))
}
}
