% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_input_info}
\alias{onnx_input_info}
\title{Get Input Information}
\usage{
onnx_input_info(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
A list of TensorInfo objects containing input tensor metadata
}
\description{
Retrieve information about model input tensors.
}
\examples{
if (check_onnx_runtime_available()) {
  session <- onnx_example_session()
  input_info <- onnx_input_info(session)
  print(input_info)
}
}
