% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_model_path}
\alias{onnx_model_path}
\title{Get Model Path}
\usage{
onnx_model_path(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
Character string with the model path
}
\description{
Get the model path from a session.
}
\examples{
if (check_onnx_runtime_available()) {
  session <- onnx_example_session()
  model_path <- onnx_model_path(session)
  cat("Model path:", model_path, "\n")
}
}
