% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_model_zoo_models}
\alias{onnx_model_zoo_models}
\title{ONNX Model Zoo Models}
\usage{
onnx_model_zoo_models()
}
\value{
A character vector of available models
}
\description{
List available models from the ONNX Model Zoo.
}
