% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_output_info}
\alias{onnx_output_info}
\title{Get Output Information}
\usage{
onnx_output_info(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
A list of TensorInfo objects containing output tensor metadata
}
\description{
Retrieve information about model output tensors.
}
\examples{
if (check_onnx_runtime_available()) {
  session <- onnx_example_session()
  output_info <- onnx_output_info(session)
  print(output_info)
}
}
