% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_run}
\alias{onnx_run}
\title{Run ONNX Inference}
\usage{
onnx_run(session, inputs)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}

\item{inputs}{A named list of input tensors. Names should match model input names.}
}
\value{
A named list of output tensors
}
\description{
Execute inference on an ONNX model with input data.
}
\examples{
if (check_onnx_runtime_available()) {
  session <- onnx_example_session()
  inputs <- list(Input3 = array(0, dim = c(1, 1, 28, 28)))
  outputs <- onnx_run(session, inputs)
}
}
