% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_onnx_runtime.R
\name{onnx_runtime_lib_path}
\alias{onnx_runtime_lib_path}
\title{Get ONNX Runtime Library Path}
\usage{
onnx_runtime_lib_path()
}
\value{
Character string with the library path.
}
\description{
Get the path to the ONNX Runtime library for the current platform.
}
\keyword{internal}
