% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{onnxruntime_install_path}
\alias{onnxruntime_install_path}
\title{A simple exported version of install_path
Returns the onnxruntime installation path.}
\usage{
onnxruntime_install_path()
}
\description{
A simple exported version of install_path
Returns the onnxruntime installation path.
}
