% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{RSession}
\alias{RSession}
\alias{$.RSession}
\title{Return RSession for onnx}
\format{
An object of class \code{environment} of length 2.
}
\usage{
RSession
}
\description{
Return RSession for onnx
}
\keyword{datasets}
