% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{$.TensorInfo}
\alias{$.TensorInfo}
\title{S3 Methods for TensorInfo and RSession}
\usage{
\method{$}{TensorInfo}(x, name)
}
\arguments{
\item{x}{A TensorInfo or RSession object}

\item{name}{Property name to access}
}
\description{
These methods provide convenient access to object properties.
}
