% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_download_model}
\alias{onnx_download_model}
\title{ONNX Download Model}
\usage{
onnx_download_model(name, url)
}
\arguments{
\item{name}{Character string specifying the model name}

\item{url}{Character string specifying the download URL}
}
\value{
Character string with the path to the downloaded model
}
\description{
Download a model from the ONNX Model Zoo.
}
