% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_download_models}
\alias{onnx_download_models}
\title{ONNX Download Models}
\usage{
onnx_download_models(subset = "vision")
}
\arguments{
\item{subset}{Character string specifying the model subset (vision, text, etc.)}
}
\value{
A named character vector of available models
}
\description{
List available models from the ONNX Model Zoo.
}
