% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_example_session}
\alias{onnx_example_session}
\title{ONNX Example Session}
\usage{
onnx_example_session(model_name = "mnist", providers = NULL)
}
\arguments{
\item{model_name}{Character string specifying the model name (default: "mnist")}

\item{providers}{Optional execution providers}
}
\value{
An RSession object
}
\description{
Create a session with an example model.
}
