% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_model_path}
\alias{onnx_model_path}
\title{Get Model Path}
\usage{
onnx_model_path(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
A character string containing the model file path
}
\description{
Retrieve the file path of the loaded model.
}
\examples{
\dontrun{
session <- onnx_session("path/to/model.onnx")
model_path <- onnx_model_path(session)
print(model_path)
}
}
