% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onnx_interface.R
\name{onnx_providers}
\alias{onnx_providers}
\title{Get Execution Providers}
\usage{
onnx_providers(session)
}
\arguments{
\item{session}{An RSession object created by onnx_session()}
}
\value{
A character vector of execution provider names
}
\description{
Retrieve information about execution providers used by the session.
}
\examples{
\dontrun{
session <- onnx_session("path/to/model.onnx")
providers <- onnx_providers(session)
print(providers)
}
}
